+++ pkgs/test/test/runner/compact_reporter_test.dart
@@ -413,10 +413,24 @@ void main() {
           +2: All tests passed!''', args: ['--run-skipped']);
     });
   });
+
+  test('Directs users to enable stack trace chaining if disabled', () async {
+    await _expectReport(
+        '''test('failure 1', () => throw TestFailure('oh no'));''', '''
+        +0: loading test.dart
+        +0: failure 1
+        +0 -1: failure 1 [E]
+          oh no
+          test.dart 6:25  main.<fn>
+        +0 -1: Some tests failed.
+        Consider enabling the flag chain-stack-traces to receive more detailed exceptions.
+        For example, 'dart test --chain-stack-traces'.''',
+        chainStackTraces: false);
+  });
 }
 
 Future<void> _expectReport(String tests, String expected,
-    {List<String> args = const []}) async {
+    {List<String> args = const [], bool chainStackTraces = true}) async {
   await d.file('test.dart', '''
     import 'dart:async';
 
@@ -427,8 +441,11 @@ $tests
     }
   ''').create();
 
-  var test = await runTest(['test.dart', '--chain-stack-traces', ...args],
-      reporter: 'compact');
+  var test = await runTest([
+    'test.dart',
+    if (chainStackTraces) '--chain-stack-traces',
+    ...args,
+  ], reporter: 'compact');
   await test.shouldExit();
 
   var stdoutLines = await test.stdout.rest.toList();