+++ pkgs/test_core/lib/src/runner/reporter/expanded.dart
@@ -81,6 +81,10 @@ class ExpandedReporter implements Reporter {
   /// Whether the reporter is paused.
   var _paused = false;
 
+  // Whether a notice should be logged about enabling stack trace chaining at
+  // the end of all tests running.
+  var _shouldPrintStackTraceChainingNotice = false;
+
   /// The set of all subscriptions to various streams.
   final _subscriptions = <StreamSubscription>{};
 
@@ -197,6 +201,11 @@ class ExpandedReporter implements Reporter {
 
   /// A callback called when [liveTest] throws [error].
   void _onError(LiveTest liveTest, error, StackTrace stackTrace) {
+    if (!liveTest.test.metadata.chainStackTraces &&
+        !liveTest.suite.isLoadSuite) {
+      _shouldPrintStackTraceChainingNotice = true;
+    }
+
     if (liveTest.state.status != Status.complete) return;
 
     _progressLine(_description(liveTest), suffix: ' $_bold$_red[E]$_noColor');
@@ -240,6 +249,14 @@ class ExpandedReporter implements Reporter {
       _progressLine('All tests skipped.');
     } else {
       _progressLine('All tests passed!');
+    }
+
+    if (_shouldPrintStackTraceChainingNotice) {
+      _sink
+        ..writeln('')
+        ..writeln('Consider enabling the flag chain-stack-traces to '
+            'receive more detailed exceptions.\n'
+            "For example, 'dart test --chain-stack-traces'.");
     }
   }